README
Welcome to the repository for the study of *Uncinaria* *stenocephala* genomics. Within this repository, we focus primarily on *U.* *stenocephala* and its genomics and or transcriptomics; however, we draw on the genomics of four hookworms (Ancylostomatidae), two members of Strongylidae and one member of Chabertidae. These seven comparative genomes were selected for their varied lifecycles and pathogenesis, as well as their close relationship with *U.* *stenocephala*. The contents of this repository as split into four sections: (1) sampling and sequencing, (2) data preparation and annotation, (3) genomics and transcriptomics and (4) pipeline development and practicing.